基础信息支持克隆

+++ MgProductBasicSvr/MgProductBasicSvr-server/src/main/java/fai/MgProductBasicSvr/domain/serviceproc/ProductProc.java
@@ -460,22 +460,6 @@ public class ProductProc {
         return refRowCount.value;
     }
 
-    public void delProduct(int aid, int pdId) {
-        int rt;
-        if(pdId <= 0) {
-            rt = Errno.ARGS_ERROR;
-            throw new MgException(rt, "args err;flow=%d;aid=%d;pdId=%s", m_flow, aid, pdId);
-        }
-
-        ParamMatcher matcher = new ParamMatcher(ProductEntity.Info.AID, ParamMatcher.EQ, aid);
-        matcher.and(ProductEntity.Info.PD_ID, ParamMatcher.EQ, pdId);
-        rt = m_dao.delete(matcher);
-        if(rt != Errno.OK) {
-            throw new MgException(rt, "del product list error;flow=%d;aid=%d;pdId=%d;", m_flow, aid, pdId);
-        }
-        Log.logStd("del product ok;aid=%d;pdId=%d;", aid, pdId);
-    }
-
     public int deleteProductList(int aid, int tid, FaiList<Integer> pdIds, boolean softDel) {
         int rt;
         if(pdIds == null || pdIds.isEmpty()) {
@@ -512,12 +496,22 @@ public class ProductProc {
             // 插入
             addSagaList(aid, list);
         }
-        Ref<Integer> refRowCount = new Ref<>();
-        rt = m_dao.delete(matcher, refRowCount);
-        if(rt != Errno.OK) {
-            throw new MgException(rt, "del product list error;flow=%d;aid=%d;pdIds=%s;", m_flow, aid, pdIds);
+
+        return delProduct(aid, matcher);
+    }
+
+    public int delProduct(int aid, ParamMatcher matcher) {
+        int rt;
+        if(matcher == null || matcher.isEmpty()) {
+            rt = Errno.ARGS_ERROR;
+            throw new MgException(rt, "args err, matcher is null;flow=%d;aid=%d;matcher=%s", m_flow, aid, matcher);
         }
 
+        Ref<Integer> refRowCount = new Ref<>();
+        rt = m_dao.delete(matcher, refRowCount);
+        if(rt != Errno.OK){
+            throw new MgException(rt, "del product list error;flow=%d;aid=%d;matcher=%s", m_flow, aid, matcher.toJson());
+        }
         return refRowCount.value;
     }
 
@@ -621,7 +615,15 @@ public class ProductProc {
         searchArg.matcher.and(ProductEntity.Info.AID, ParamMatcher.EQ, aid);
 
         Ref<FaiList<Param>> listRef = new Ref<>();
+
+        // 因为克隆可能获取其他aid的数据，所以根据传进来的aid设置tablename
+        m_dao.setTableName(aid);
+
         int rt = m_dao.select(searchArg, listRef, selectFields);
+
+        // 查完之后恢复最初的tablename
+        m_dao.restoreTableName();
+
         if(rt != Errno.OK && rt != Errno.NOT_FOUND) {
             throw new MgException(rt, "get error;flow=%d;aid=%d;", m_flow, aid);
         }
